+++ src/ham_fisted/api.clj
@@ -503,7 +503,10 @@ ham_fisted.PersistentHashMap
 (defn linked-hashmap
   "Linked hash map using clojure's equiv pathways.  This hashmap behaves slightly
   different in that if you put the same value as previously the node isn't marked
-  as modified"
+  as modified.  It has an accelerated union pathway (.union member) that will only
+  work if the other arg is also a linked hashmap.  Union will not change the order
+  of the original hashmap but new keys may be added and will appear later in iteration
+  order."
   (^ham_fisted.LinkedHashMap [] (ham_fisted.LinkedHashMap.))
   (^ham_fisted.LinkedHashMap [data]
    (let [rv (ham_fisted.LinkedHashMap.)]
@@ -2316,6 +2319,32 @@ ham-fisted.api> (binary-search data 1.1 nil)
   ^long [f data]
   @(reduce hamf-rf/consumer-accumulator (MinIdx. Long/MAX_VALUE 0 -1 (ensure-obj->long f))
            data))
+
+
+(defn intersect-sets
+  "Given a sequence of sets, efficiently perform the intersection of them.  This algorithm is usually faster and has a more stable
+   runtime than (reduce clojure.set/intersection sets) which degrades depending on the order of the sets and the pairwise
+   intersection of the initial sets."
+  [sets]
+  (let [sets (vec sets)
+        ns (count sets)]
+    (case ns
+      0 #{}
+      1 (sets 0)
+      (let [min-idx (mmin-idx (fn ^long [arg] (.size ^Set arg)) sets)]
+        (-> (reduce (fn [^Set rv val]
+                      (when (reduce (fn [rv ^long idx]
+                                      (if (== idx min-idx)
+                                        rv
+                                        (if-not (.contains ^Set (sets idx) val)
+                                          (reduced false)
+                                          true)))
+                                    true
+                                    (range ns))
+                        (.add rv val))
+                      rv)
+                    (mut-set)
+                    (sets min-idx)))))))