+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.009
+ * new linked hashmap implementation with equiv-semantics and fast union op.
+
+# 1.008
+ * Fast set intersection for longer sequencers of sets (intersect-sets).
+
 # 1.007
  * Fast pathways for finding min/max index of a collection of objects in a similar way to min-key and max-key.